Fix samples.
