+++ build.sh
@@ -1,1 +1,2 @@
 coffee -c --output lib src
+mv lib/v1sdk.js .
