--- index.js
@@ -1,8 +1,0 @@
-// Generated by CoffeeScript 1.6.3
-(function() {
-  module.exports = {
-    V1Meta: require('./v1meta').V1Meta,
-    V1Server: require('./client').V1Server
-  };
-
-}).call(this);
