+++ lib/query.js
@@ -1,0 +1,19 @@
+// Generated by CoffeeScript 1.7.1
+(function() {
+  var V1Query;
+
+  V1Query = (function() {
+    function V1Query(v1instance, assettype, options) {
+      this.v1instance = v1instance;
+      this.assettype = assettype;
+      this.options = options != null ? options : {};
+      void 0;
+    }
+
+    V1Query.prototype.execute = function() {};
+
+    return V1Query;
+
+  })();
+
+}).call(this);
