+++ lib/test.js
@@ -1,0 +1,26 @@
+// Generated by CoffeeScript 1.7.1
+(function() {
+  var client, print_errors, print_xml, server;
+
+  client = require('./client');
+
+  server = new client.V1Server();
+
+  print_errors = function(error) {
+    return console.log(error);
+  };
+
+  print_xml = function(xml) {
+    return console.log(xml);
+  };
+
+  server.get_meta_xml({
+    asset_type_name: 'Story'
+  }, console.log, console.log);
+
+  server.get_asset_xml({
+    asset_type_name: 'Story',
+    id: 1005
+  }, console.log, console.log);
+
+}).call(this);
