+++ lib/test_all_types.js
@@ -1,0 +1,21 @@
+// Generated by CoffeeScript 1.7.1
+(function() {
+  var V1Meta, v1;
+
+  V1Meta = require('./v1meta').V1Meta;
+
+  v1 = new V1Meta();
+
+  v1.for_all_types(function(err, types) {
+    var i, type, _i, _len, _results;
+    _results = [];
+    for (_i = 0, _len = types.length; _i < _len; _i++) {
+      type = types[_i];
+      i = new type('id');
+      console.log(i);
+      _results.push(console.log(i._v1_ops));
+    }
+    return _results;
+  });
+
+}).call(this);
