+++ lib/test_meta.js
@@ -1,0 +1,72 @@
+// Generated by CoffeeScript 1.7.1
+(function() {
+  var V1Meta, V1Server, server, v1;
+
+  V1Meta = require('./v1meta').V1Meta;
+
+  V1Server = require('./client').V1Server;
+
+  server = new V1Server();
+
+  v1 = new V1Meta(server);
+
+  v1.query({
+    select: ['Email', 'Username', 'ID'],
+    from: "Member",
+    where: {
+      ID: 'Member:1000',
+      Email: 'andre@company.com',
+      Username: 'andre'
+    },
+    success: function(result) {
+      var message, output;
+      message = 'Email: ' + result.Email + "<br/>\n" + 'Username: ' + result.Username + "<br/>\n" + 'ID: ' + result.ID;
+      output = function() {
+        if (server.useBrowserHttpStack) {
+          return document.getElementById("query").innerHTML += "<b>Query result:</b><br/><br/>" + message;
+        } else {
+          return console.log("\nQuery result:\n\n" + message);
+        }
+      };
+      return setTimeout(output, 2000);
+    },
+    error: function(err) {
+      return console.log("Error: " + err);
+    }
+  });
+
+  v1.trans_query({
+    from: "Member",
+    where: {
+      ID: 'Member:1000',
+      Email: 'andre@company.com',
+      Username: 'andre'
+    },
+    select: ["Email", "ID", "Username"],
+    success: function(trans) {
+      var message, output;
+      message = "";
+      trans.iter(function(member) {
+        if (server.useBrowserHttpStack) {
+          return message += member + "<br/>\n";
+        } else {
+          return message += member + "\n";
+        }
+      });
+      output = function() {
+        if (server.useBrowserHttpStack) {
+          return document.getElementById("trans_query").innerHTML += "<b>Transaction Query result:</b><br/><br/>" + message;
+        } else {
+          return console.log("\nTransaction Query result:\n\n" + message);
+        }
+      };
+      return setTimeout(output, 2000);
+    },
+    error: function(err) {
+      return console.log("Query failed:\n" + util.inspect(err));
+    }
+  });
+
+  "v1.with_poller (poller) ->\n    poller.on_new_asset 'Story',\n        'CreateDate Moment Owners',\n        (story) ->\n            console.log story\n            story.Touched = now()\n    poller.commit and_log_callback";
+
+}).call(this);
