+++ samples/sample2.js
@@ -1,6 +1,6 @@
-var  v1sdk = require('./v1sdk');
+var  v1sdk = require('../v1sdk');
 
-var hostname = "www14.v1host.com"; 
+var hostname = "www14.v1host.com";
 var instance = "v1sdktesting";
 var username = "api";
 var password = "api";
@@ -21,7 +21,7 @@ v1.query({
         console.log(result.Email);
         console.log(result.Username);
         console.log(result.ID);
-    },    
+    },
     error: function(err) { // NOTE: this is not working correctly yet, not called...
         console.log(err);
     }
