+++ src/v1sdk.coffee
@@ -1,3 +1,3 @@
 module.exports =
-	V1Meta: require('./v1meta').V1Meta
-	V1Server: require('./client').V1Server
+	V1Meta: require('./lib/v1meta').V1Meta
+	V1Server: require('./lib/client').V1Server
