+++ v1sdk.js
@@ -1,0 +1,8 @@
+// Generated by CoffeeScript 1.7.1
+(function() {
+  module.exports = {
+    V1Meta: require('./lib/v1meta').V1Meta,
+    V1Server: require('./lib/client').V1Server
+  };
+
+}).call(this);
